+++ src/Components/ChatItem.module.css
@@ -3,7 +3,6 @@
   flex-direction: row-reverse;
   background-color: pink;
 }
-
 .my__chat {
   justify-content: flex-end;
   background-color: #01ff00;
@@ -11,7 +10,6 @@
 .girl__chat_wrap {
   justify-content: flex-start;
 }
-
 .my__chat_wrap {
   justify-content: flex-end;
 }
