+++ src/Components/index.js
@@ -2,3 +2,4 @@ export { default as AnswersList } from "./AnswersList";
 export { default as AnswerItem } from "./AnswerItem";
 export { default as ChatList } from "./ChatList";
 export { default as ChatItem } from "./ChatItem";
+export { default as LoadingMotion } from "./LoadingMotion";
