+++ src/Pages/index.js
@@ -2,4 +2,3 @@ export { default as Home } from "./Home";
 export { default as GirlsSelect } from "./GirlsSelect";
 export { default as StartChat } from "./StartChat";
 export { default as ChatRoom } from "./ChatRoom";
-export { default as LoadingMotion} from "./LoadingMotion";
